webscraping/requirements.txt
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==6.7
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-spreadsheet==0.0.6
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.18.1
idna==2.7
isort==4.3.21
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==0.23
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.0
Pillow==7.2.0
plaid-python==5.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pygsheets==2.0.3.1
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.20.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.3
toml==0.10.1
tqdm==4.48.2
uritemplate==3.0.1
urllib3==1.24.3
Werkzeug==1.0.1
wrapt==1.12.1
